Refactor notifications module to enhance organization and functionality
- Consolidated notification services and processors into dedicated directories for improved structure and clarity.
- Updated import paths and introduced index files for streamlined imports across the module.
- Implemented a NotificationQueueService to manage notification queuing for various channels, enhancing modularity.
- Added new services for email, Slack, Telegram, and webhook notifications, ensuring better separation of concerns.
- Maintained existing functionality while improving code maintainability and readability.
