--- src/notifications/REFACTOR_PLAN.md
@@ -1,86 +1,0 @@
-# Notifications Module Refactor Plan - Simple File Organization
-
-## 🎯 **Objective**
-
-Organize existing notification module files into logical folders (/services and /processors) to improve code organization without changing functionality.
-
-## 📁 **New Structure (Completed)**
-
-```
-notifications/
-├── notifications.service.ts           # Main orchestrator
-├── notifications.controller.ts        # API endpoints
-├── notifications.module.ts            # Module configuration
-├── notifications.md                   # Documentation
-├── notifications.service.spec.ts      # Tests
-├── services/                          # ✅ All notification services
-│   ├── index.ts                       # Export barrel
-│   ├── email-notification.service.ts  # Email notifications
-│   ├── slack-notification.service.ts  # Slack notifications
-│   ├── telegram-notification.service.ts # Telegram notifications
-│   └── webhook-notification.service.ts # Webhook notifications
-└── processors/                        # ✅ All queue processors
-    ├── index.ts                       # Export barrel
-    ├── alerts.processor.ts            # Main alert processor
-    ├── email-notification.processor.ts # Email queue processor
-    ├── slack-notification.processor.ts # Slack queue processor
-    ├── telegram-notification.processor.ts # Telegram queue processor
-    └── webhook-notification.processor.ts # Webhook queue processor
-```
-
-## ✅ **Phase 1: File Organization - COMPLETED**
-
-### **What was done:**
-
-1. **Created `/services` folder** - Organized all notification service files
-2. **Created `/processors` folder** - Organized all queue processor files
-3. **Updated import paths** - Updated notifications.module.ts and notifications.service.ts
-4. **Created index.ts files** - Export barrels for clean imports
-5. **Cleaned up old files** - Removed original files from root directory
-
-### **Files moved:**
-
-- `notif.email.service.ts` → `services/email-notification.service.ts`
-- `notif.slack.service.ts` → `services/slack-notification.service.ts`
-- `notif.telegram.service.ts` → `services/telegram-notification.service.ts`
-- `notif.webhook.service.ts` → `services/webhook-notification.service.ts`
-- `notif.email.processor.ts` → `processors/email-notification.processor.ts`
-- `notif.slack.processor.ts` → `processors/slack-notification.processor.ts`
-- `notif.telegram.processor.ts` → `processors/telegram-notification.processor.ts`
-- `notif.webhook.processor.ts` → `processors/webhook-notification.processor.ts`
-- `alerts.processor.ts` → `processors/alerts.processor.ts`
-
-### **Key Benefits Achieved:**
-
-- ✅ **Better file organization** - Clear separation between services and processors
-- ✅ **Consistent naming** - Removed `notif.` prefix, used descriptive names
-- ✅ **Clean imports** - Export barrels for easy importing
-- ✅ **Maintained functionality** - No changes to existing logic
-- ✅ **No breaking changes** - All existing functionality preserved
-
-## 📊 **Results**
-
-### **Before:**
-
-- 13 files in root notifications directory
-- Inconsistent naming (`notif.*.service.ts`, `notif.*.processor.ts`)
-- Mixed concerns in same directory
-
-### **After:**
-
-- 5 files in root directory (core module files)
-- 4 organized services in `/services` folder
-- 5 organized processors in `/processors` folder
-- Consistent, descriptive naming
-- Clear separation of concerns
-
-## 🎯 **Success Metrics**
-
-- ✅ **File organization improved** - Clear logical grouping
-- ✅ **Developer experience enhanced** - Easier to find and understand code
-- ✅ **No functionality lost** - All existing features work unchanged
-- ✅ **Clean structure** - Better foundation for future improvements
-
----
-
-**Phase 1 Complete!** The notifications module now has a clean, organized structure while maintaining all existing functionality. The module is ready for future improvements if needed.
