+++ src/notifications/processors/index.ts
@@ -1,5 +1,5 @@
 export * from './alerts.processor';
-export * from './email-notification.processor';
-export * from './slack-notification.processor';
-export * from './telegram-notification.processor';
-export * from './webhook-notification.processor';
+export * from './email.processor';
+export * from './slack.processor';
+export * from './telegram.processor';
+export * from './webhook.processor';
