+++ src/notifications/processors/webhook.processor.ts
@@ -6,9 +6,9 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Alert } from 'src/alerts/entities/alert.entity';
 import { UserContract } from 'src/user-contracts/entities/user-contract.entity';
-import { EmailNotificationService } from '../services/email-notification.service';
+import { WebhookNotificationService } from '../services/webhook.service';
 
-interface EmailNotificationData {
+interface WebhookNotificationData {
   alertId: string;
   alertType: AlertType;
   destination: string;
@@ -16,12 +16,12 @@ interface EmailNotificationData {
 }
 
 @Injectable()
-@Processor('notif-email')
-export class EmailNotificationProcessor extends WorkerHost {
-  private readonly logger = new Logger(EmailNotificationProcessor.name);
+@Processor('notif-webhook')
+export class WebhookNotificationProcessor extends WorkerHost {
+  private readonly logger = new Logger(WebhookNotificationProcessor.name);
 
   constructor(
-    private readonly emailService: EmailNotificationService,
+    private readonly webhookService: WebhookNotificationService,
     @InjectRepository(Alert)
     private alertsRepository: Repository<Alert>,
     @InjectRepository(UserContract)
@@ -30,15 +30,17 @@ export class EmailNotificationProcessor extends WorkerHost {
     super();
   }
 
-  async process(job: Job<EmailNotificationData, void, string>): Promise<void> {
+  async process(
+    job: Job<WebhookNotificationData, void, string>,
+  ): Promise<void> {
     const { alertId, alertType, destination, userId } = job.data;
 
     // Log notification details
-    this.logger.log(`Processing Email notification for alert: ${alertId}`);
+    this.logger.log(`Processing Webhook notification for alert: ${alertId}`);
     this.logger.log(`Attempt number: ${job.attemptsMade + 1}`);
     this.logger.log(`Alert type: ${alertType}`);
     this.logger.log(`User ID: ${userId}`);
-    this.logger.log(`Email address: ${destination}`);
+    this.logger.log(`Webhook URL: ${destination}`);
 
     try {
       // Fetch the alert details to get more context
@@ -59,13 +61,10 @@ export class EmailNotificationProcessor extends WorkerHost {
         contractName = alert.userContract.name || 'Unknown contract';
       }
 
-      // Create recipient name from user data or use a default
-      const recipientName = alert.user?.name || 'Stylus User';
-
-      // Use the email service to send the notification
-      await this.emailService.sendNotification({
+      // Use the webhook service to send the notification
+      await this.webhookService.sendNotification({
         destination,
-        recipientName,
+        alertId,
         alertType,
         value: alert.value,
         contractName,
@@ -73,17 +72,21 @@ export class EmailNotificationProcessor extends WorkerHost {
         triggeredCount: alert.triggeredCount,
       });
 
-      this.logger.log(`Email notification sent successfully to ${destination}`);
+      this.logger.log(
+        `Webhook notification sent successfully to ${destination}`,
+      );
 
       // Update job progress to indicate completion
       await job.updateProgress(100);
     } catch (error) {
       if (error instanceof Error) {
         this.logger.error(
-          `Error processing email notification: ${error.message}`,
+          `Error processing Webhook notification: ${error.message}`,
         );
       } else {
-        this.logger.error(`Error processing email notification: Unknown error`);
+        this.logger.error(
+          `Error processing Webhook notification: Unknown error`,
+        );
       }
       throw error; // Re-throw to let BullMQ handle retries
     }
