+++ src/notifications/services/index.ts
@@ -1,4 +1,7 @@
-export * from './email-notification.service';
-export * from './slack-notification.service';
-export * from './telegram-notification.service';
-export * from './webhook-notification.service';
+export * from './email.service';
+export * from './slack.service';
+export * from './telegram.service';
+export * from './webhook.service';
+export * from './queue.service';
+export * from './mock.service';
+export * from './timing.service';
