+++ src/notifications/services/mock.service.ts
@@ -1,0 +1,71 @@
+import { Injectable, Logger } from '@nestjs/common';
+import { AlertType } from 'src/alerts/entities/alert.entity';
+import { User } from 'src/users/entities/user.entity';
+import { WebhookNotificationService } from './webhook.service';
+import { SlackNotificationService } from './slack.service';
+import { TelegramNotificationService } from './telegram.service';
+import { EmailNotificationService } from './email.service';
+
+@Injectable()
+export class MockNotificationService {
+  private readonly logger = new Logger(MockNotificationService.name);
+
+  constructor(
+    private webhookService: WebhookNotificationService,
+    private slackService: SlackNotificationService,
+    private telegramService: TelegramNotificationService,
+    private emailService: EmailNotificationService,
+  ) {}
+
+  async sendMockNotification(
+    user: User,
+    notificationChannel: 'webhook' | 'slack' | 'telegram' | 'email',
+  ) {
+    const mockData = {
+      alertId: '123',
+      alertValue: '100',
+      alertContractAddress: '0x123',
+      alertContractName: 'Mock Contract',
+      alertType: AlertType.EVICTION,
+      userId: user.id,
+      destination: 'test@test.com',
+      alertTypeDisplay: 'Mock Alert Type',
+    };
+
+    const destination =
+      user.alertsSettings[`${notificationChannel}Settings`]?.destination;
+
+    if (!destination) {
+      this.logger.warn(`No destination found for ${notificationChannel}`);
+      return;
+    }
+
+    this.logger.log(
+      `Sending mock ${notificationChannel} notification to ${destination}`,
+    );
+
+    const sendServices = {
+      webhook: this.webhookService,
+      slack: this.slackService,
+      telegram: this.telegramService,
+      email: this.emailService,
+      default: () => {},
+    };
+
+    await sendServices[notificationChannel].sendNotification({
+      destination,
+      recipientName: user.name,
+      alertId: mockData.alertId,
+      alertType: mockData.alertType,
+      value: mockData.alertValue,
+      contractName: mockData.alertContractName,
+      contractAddress: mockData.alertContractAddress,
+      triggeredCount: 0,
+    });
+
+    return {
+      success: true,
+      message: 'Mock notification sent',
+    };
+  }
+}
